Added a "click zone" to the søckets
Søckets are fairly small and hard to click.
To make them more clickable I added a transparent square over them.
This increases the clickable area of the søckets.
The click zone element also functions as a hover indicator.
It displays a small solid circle inside the søcket ring when hovered.
This is accomplished using a radial gradient.
The radial gradient is defined by svgDefs.js
It is added by N0deView to the top of the SVG.
It is then referenced by ID from the CSS.

Other changes:
- Moved no-select to top area of CSS.
- Moved various style attributes from JS to CSS.
- Added a classname to the nøde frame.
- Broke apart draw functions of S0cket
 > so class can manage individual elements